Fixed datatype on add items on InventoryManager.java Remove ln on print staement on menuManager.java
added items.tx
, remove data request.txt

+++ src/InventoryManager/InventoryManager.java
@@ -20,7 +20,7 @@ public class InventoryManager {
         String itemName = sc.nextLine();
 
         System.out.print("Enter Price: ");
-        int itemPrice = sc.nextInt();
+        double itemPrice = sc.nextDouble();
         sc.nextLine();
 
         System.out.print("Enter Item Quantity: ");
